chore: refactor ortelius helper functions
