+++ src/Explorer/index.ts
@@ -1,3 +1,2 @@
-export * from './utils';
 export * from './snowtrace';
 export * from './ortelius';
