+++ src/Explorer/ortelius/index.ts
@@ -1,2 +1,3 @@
 export * from './requests';
 export * from './types';
+export * from './utils';
