+++ src/Explorer/ortelius/requests.ts
@@ -150,3 +150,39 @@ export async function getAddressHistory(
         return [...acc, ...txs];
     }, []);
 }
+
+/**
+ * Given an array of addresses, checks which chain each address was already used on
+ * @param addrs
+ */
+export async function getAddressChains(addrs: string[]) {
+    if (!explorer_api) {
+        throw NO_EXPLORER_API;
+    }
+
+    // Strip the prefix
+    let rawAddrs = addrs.map((addr) => {
+        return addr.split('-')[1];
+    });
+
+    let urlRoot = `/v2/addressChains`;
+
+    let res = await explorer_api.post(urlRoot, {
+        address: rawAddrs,
+        disableCount: ['1'],
+    });
+
+    return res.data.addressChains;
+}
+
+export async function getAddressDetailX(addr: string) {
+    if (!explorer_api) {
+        throw NO_EXPLORER_API;
+    }
+
+    let addrRaw = addr.split('-')[1];
+    let url = `/x/addresses/${addrRaw}`;
+
+    let res = await explorer_api.get(url);
+    return res.data;
+}
