+++ src/History/base_tx_parser.ts
@@ -10,7 +10,7 @@ import * as Assets from '@/Asset/Assets';
 import { bnToLocaleString, getTxFeeX } from '@/utils';
 import { AVMConstants } from 'avalanche/dist/apis/avm';
 import { BN } from 'avalanche';
-import { getAssetBalanceFromUTXOs, getNFTBalanceFromUTXOs, parseMemo } from '@/History/history_helpers';
+import { getNFTBalanceFromUTXOs, parseMemo } from '@/History/history_helpers';
 import {
     filterDuplicateStrings,
     getAssetOutputs,
