+++ src/index.ts
@@ -1,11 +1,11 @@
 export * as Network from '@/Network';
 export * as Assets from '@/Asset';
 export * as Common from './common';
-export * as History from '@/History';
 
 export * from './types';
 export * from './Wallet';
 export * from './Explorer';
+export * from './History';
 
 import * as NetworkConstants from '@/Network/constants';
 import * as Utils from '@/utils';
